Shodai - first published app
Any advice on doing this better is greatly appreciated. I had several HTML5 elements in the index.html that the Chrome Dev Tools editor just "errored" me to death on for days, until I finally gave up. The end result is not nearly as elegant as I would have prefered for such a simple app.